removed adapter, moved logic into model
